requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.7.0
    # via pydantic
anyio==4.5.2
    # via
    #   httpx
    #   openai
certifi==2024.12.14
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
click==8.1.8
    # via nltk
diffusers==0.32.0
    # via -r requirements.in
distro==1.9.0
    # via openai
exceptiongroup==1.2.2
    # via anyio
filelock==3.16.1
    # via
    #   diffusers
    #   huggingface-hub
    #   torch
    #   triton
fsspec==2024.12.0
    # via
    #   huggingface-hub
    #   torch
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.28.1
    # via openai
huggingface-hub==0.27.0
    # via diffusers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==8.5.0
    # via diffusers
jinja2==3.1.5
    # via torch
jiter==0.8.2
    # via openai
joblib==1.4.2
    # via nltk
markupsafe==2.1.5
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.1
    # via torch
nltk==3.9.1
    # via -r requirements.in
numpy==1.24.4
    # via
    #   diffusers
    #   pandas
nvidia-cublas-cu12==12.1.3.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.1.105
    # via torch
nvidia-cuda-nvrtc-cu12==12.1.105
    # via torch
nvidia-cuda-runtime-cu12==12.1.105
    # via torch
nvidia-cudnn-cu12==9.1.0.70
    # via torch
nvidia-cufft-cu12==11.0.2.54
    # via torch
nvidia-curand-cu12==10.3.2.106
    # via torch
nvidia-cusolver-cu12==11.4.5.107
    # via torch
nvidia-cusparse-cu12==12.1.0.106
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-nccl-cu12==2.20.5
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
nvidia-nvtx-cu12==12.1.105
    # via torch
openai==1.58.1
    # via -r requirements.in
packaging==24.2
    # via huggingface-hub
pandas==2.0.3
    # via -r requirements.in
pillow==10.4.0
    # via diffusers
pydantic==2.10.4
    # via openai
pydantic-core==2.27.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.2
    # via pandas
pyyaml==6.0.2
    # via huggingface-hub
regex==2024.11.6
    # via
    #   diffusers
    #   nltk
requests==2.32.3
    # via
    #   -r requirements.in
    #   diffusers
    #   huggingface-hub
safetensors==0.4.5
    # via diffusers
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sympy==1.13.3
    # via torch
torch==2.4.1
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   nltk
    #   openai
triton==3.0.0
    # via torch
typing-extensions==4.12.2
    # via
    #   annotated-types
    #   anyio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   torch
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
zipp==3.20.2
    # via importlib-metadata
